refactor: ID, 생성 시각, 수정 시각 공통 로직 BaseEntity 추상 클래스로 분리
