--- src/main/java/qna/domain/TimeAudit.java
@@ -1,26 +1,0 @@
-package qna.domain;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-@Embeddable
-public class TimeAudit {
-
-    @Column(name = "created_at", nullable = false)
-    @CreatedDate
-    private LocalDateTime createdAt = LocalDateTime.now();
-
-    @Column(name = "updated_at")
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-
-    protected TimeAudit() {
-    }
-
-    public void renewUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
